Reorganize themes/*.dart
