--- lib/src/themes/yaru_light.dart
@@ -1,20 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:yaru/src/colors/yaru_colors.dart';
-import 'package:yaru/src/themes/common_themes.dart';
-
-const _primaryColor = YaruColors.ubuntuOrange;
-
-final _lightColorScheme = ColorScheme.fromSwatch(
-  primarySwatch: _primaryColor,
-  primaryColorDark: YaruColors.coolGrey,
-  accentColor: _primaryColor,
-  cardColor: Colors.white,
-  backgroundColor: YaruColors.porcelain,
-  errorColor: YaruColors.red,
-  brightness: Brightness.light,
-);
-
-final yaruLight = createYaruLightTheme(
-    colorScheme: _lightColorScheme,
-    primaryColor: _primaryColor,
-    elevatedButtonColor: YaruColors.green);
