--- lib/src/themes/yaru_lubuntu_dark.dart
@@ -1,19 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:yaru/src/colors/flavor_colors.dart';
-import 'package:yaru/src/colors/yaru_colors.dart';
-import 'package:yaru/src/themes/common_themes.dart';
-
-const _primaryColor = FlavorColors.lubuntuBlue;
-
-final _darkColorScheme = ColorScheme.fromSwatch(
-  primarySwatch: _primaryColor,
-  primaryColorDark: YaruColors.coolGrey,
-  accentColor: _primaryColor,
-  cardColor: YaruColors.jet,
-  backgroundColor: YaruColors.jet,
-  errorColor: YaruColors.red,
-  brightness: Brightness.dark,
-);
-
-final yaruLubuntuDark = createYaruDarkTheme(
-    colorScheme: _darkColorScheme, primaryColor: _primaryColor);
