+++ lib/yaru.dart
@@ -4,19 +4,12 @@ export 'package:yaru/src/colors/flavor_colors.dart';
 export 'package:yaru/src/colors/yaru_colors.dart';
 export 'package:yaru/src/themes/common_themes.dart';
 export 'package:yaru/src/themes/high_contrast.dart';
-export 'package:yaru/src/themes/yaru_dark.dart';
-export 'package:yaru/src/themes/yaru_kubuntu_dark.dart';
-export 'package:yaru/src/themes/yaru_kubuntu_light.dart';
-export 'package:yaru/src/themes/yaru_light.dart';
-export 'package:yaru/src/themes/yaru_lubuntu_dark.dart';
-export 'package:yaru/src/themes/yaru_lubuntu_light.dart';
-export 'package:yaru/src/themes/yaru_mate_dark.dart';
-export 'package:yaru/src/themes/yaru_mate_light.dart';
-export 'package:yaru/src/themes/yaru_ubuntu_budgie_dark.dart';
-export 'package:yaru/src/themes/yaru_ubuntu_budgie_light.dart';
-export 'package:yaru/src/themes/yaru_ubuntu_studio_dark.dart';
-export 'package:yaru/src/themes/yaru_ubuntu_studio_light.dart';
-export 'package:yaru/src/themes/yaru_variants.dart';
-export 'package:yaru/src/themes/yaru_xubuntu_dark.dart';
-export 'package:yaru/src/themes/yaru_xubuntu_light.dart';
+export 'package:yaru/src/themes/kubuntu.dart';
+export 'package:yaru/src/themes/lubuntu.dart';
+export 'package:yaru/src/themes/ubuntu_budgie.dart';
+export 'package:yaru/src/themes/ubuntu_mate.dart';
+export 'package:yaru/src/themes/ubuntu_studio.dart';
+export 'package:yaru/src/themes/variant.dart';
+export 'package:yaru/src/themes/xubuntu.dart';
+export 'package:yaru/src/themes/yaru.dart';
 export 'package:yaru/src/widgets/inherited_theme.dart';
